# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UCRS API Documentation",
        "description": "University Course Registration System - Complete API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/enrollments": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "List all enrollments",
                "description": "Get paginated list of all enrollments (Admin only)",
                "operationId": "46f86247fae90db42c50de2b294234ad",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrollments retrieved"
                    },
                    "403": {
                        "description": "Forbidden (Admin only)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/courses/{courseId}/enroll": {
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Enroll in course",
                "description": "Enroll current student in a course",
                "operationId": "003291b17abf2852e9061831a3ab7701",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Enrolled successfully"
                    },
                    "422": {
                        "description": "Already enrolled"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/enrollments/my-courses": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get my enrolled courses",
                "description": "Get courses that current student is enrolled in",
                "operationId": "4cb11a8fb2a8f685abb5c30105835dc2",
                "responses": {
                    "200": {
                        "description": "My courses retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/enrollments/{id}": {
            "put": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Update enrollment",
                "description": "Update enrollment final grade (Admin only)",
                "operationId": "a7025556bd0fa80421bec4f504c04962",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "final_grade"
                                ],
                                "properties": {
                                    "final_grade": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 85.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grade updated"
                    },
                    "404": {
                        "description": "Enrollment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Drop course",
                "description": "Student drops a course enrollment",
                "operationId": "7f6eede96ccbc31ddf0f9cbb9410466a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrollment deleted"
                    },
                    "403": {
                        "description": "Not your enrollment"
                    },
                    "404": {
                        "description": "Enrollment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}